feat(crossasset)!: convert model from float64 to float32
Convert the entire crossasset package (~2300 lines) from float64 to
float32 for model weights, forward/backward pass, and training data.
This enables GPU acceleration (GPU engines work with float32) and
halves memory bandwidth, addressing issues #381 and #384 about
crossasset training being 10-100x slower than PyTorch.

BREAKING CHANGE: Public API signatures changed:
- Forward([][]float64) -> Forward([][]float32)
- Train([][][]float64, ...) -> Train([][][]float32, ...)
- TrainGPU([][][]float64, ...) -> TrainGPU([][][]float32, ...)
- Predict([][]float64) -> Predict([][]float32)
- AttentionWeights([][]float64) -> AttentionWeights([][]float32)

Changes:
- cpuEngine: NewCPUEngine[float64] -> NewCPUEngine[float32]
- All weight arrays (headW, headB, layer weights): float64 -> float32
- All layer node types: Linear[float64] -> Linear[float32], etc.
- All tensor operations: tensor.New[float64] -> tensor.New[float32]
- Serialize format bumped to v2 (4 bytes/weight instead of 8)
- AdamW optimizer: AdamW[float64] -> AdamW[float32]
- timeseries/crossasset_engine.go: converts float64 windows to float32
  at the adapter boundary
- Test tolerances relaxed for float32 precision (gradient parity,
  attention weight sums)

Config fields (LearningRate, DropoutRate) and summary metrics
(TrainResult.Losses, FinalAccuracy) remain float64.

+++ crossasset/gpu_train.go
@@ -22,7 +22,7 @@ type TrainResult struct {
 //
 // The engine parameter is accepted for API compatibility but not used.
 // Training accuracy matches PyTorch (~75% on COIN walk-forward validation).
-func (m *Model) TrainGPU(data [][][]float64, labels [][]int, tc TrainConfig,
+func (m *Model) TrainGPU(data [][][]float32, labels [][]int, tc TrainConfig,
 	_ compute.Engine[float32]) (*TrainResult, error) {
 
 	if len(data) == 0 {
@@ -54,13 +54,13 @@ func (m *Model) TrainGPU(data [][][]float64, labels [][]int, tc TrainConfig,
 			continue
 		}
 		for s := range ns {
-			logits := make([]float64, 3)
+			logits := make([]float32, 3)
 			matVecMul(logits, m.headW, outputs[s], m.config.DModel, 3)
 			vecAdd(logits, m.headB)
 			probs := softmax(logits)
 			target := labels[i][s]
 			if target >= 0 && target < 3 {
-				p := probs[target]
+				p := float64(probs[target])
 				if p < 1e-15 {
 					p = 1e-15
 				}
